Updates code to support newest Android permissions model
This commit needs some UX updates before it should be merged/pushed to
production, but the app does function. Essentially it just requests
needed permissions from the user but doesn't tell them why they are
needed.

There may be other areas of the app that need updates with regards to
permissions, this commit only gets as far as allowing the app to start